+++ external-crates/move/crates/move-compiler/src/linters/mod.rs
@@ -5,9 +5,9 @@ use move_symbol_pool::Symbol;
 
 use crate::{
     command_line::compiler::Visitor, diagnostics::codes::WarningFilter,
-    linters::shift_overflow::ShiftOperationOverflow, typing::visitor::TypingVisitor,
+    linters::redundant_assert::AssertTrueFals, typing::visitor::TypingVisitor,
 };
-pub mod shift_overflow;
+pub mod redundant_assert;
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum LintLevel {
     // No linters
@@ -20,12 +20,12 @@ pub enum LintLevel {
 
 pub const ALLOW_ATTR_CATEGORY: &str = "lint";
 pub const LINT_WARNING_PREFIX: &str = "Lint ";
-pub const SHILF_OVERFLOW_FILTER_NAME: &str = "shift_overflow";
+pub const REDUNDANT_ASSERT_FILTER_NAME: &str = "shift_overflow";
 
 pub const LINTER_DEFAULT_DIAG_CODE: u8 = 1;
 
 pub enum LinterDiagCategory {
-    ShiftOperationOverflow,
+    AssertTrueFals,
 }
 
 pub fn known_filters() -> (Option<Symbol>, Vec<WarningFilter>) {
@@ -33,9 +33,9 @@ pub fn known_filters() -> (Option<Symbol>, Vec<WarningFilter>) {
         Some(ALLOW_ATTR_CATEGORY.into()),
         vec![WarningFilter::code(
             Some(LINT_WARNING_PREFIX),
-            LinterDiagCategory::ShiftOperationOverflow as u8,
+            LinterDiagCategory::AssertTrueFals as u8,
             LINTER_DEFAULT_DIAG_CODE,
-            Some(SHILF_OVERFLOW_FILTER_NAME),
+            Some(REDUNDANT_ASSERT_FILTER_NAME),
         )],
     )
 }
@@ -44,9 +44,7 @@ pub fn linter_visitors(level: LintLevel) -> Vec<Visitor> {
     match level {
         LintLevel::None => vec![],
         LintLevel::Default | LintLevel::All => {
-            vec![shift_overflow::ShiftOperationOverflow::visitor(
-                ShiftOperationOverflow,
-            )]
+            vec![redundant_assert::AssertTrueFals::visitor(AssertTrueFals)]
         }
     }
 }
